Refactor ncf
